+++ apps/web/routes/web.php
@@ -1,7 +1,6 @@
 <?php
 
+use App\Http\Controllers\Homepage\ShowHomepageController;
 use Illuminate\Support\Facades\Route;
 
-Route::get('/', function () {
-    return view('welcome');
-})->name('home');
+Route::get('/', ShowHomepageController::class)->name('home');